Add DhcpClient. Remove dependency on odhcploc
Add a class that acts as a DHCP client.
Useful for obtaining some info from the DHCP server such as
what the domain name used inside the network is.

Also use this to detect the presence of other DHCP servers
inside the network if user selected standalone DHCP server mode.
Removes dependency on odhcploc.