Force reconfig of existing slapd
When the user completes the wizard during first use, we install
LDAP server software.
Access to the backend is secured with an admin password chosen
randomly and stored in settings.json

In the unlikely event that the settings file goes missing, we no
longer know the password to access the LDAP server still installed.

- Upon completing the wizard again, force reconfiguration if slapd
  is already installed, so that the password (and any other data) is reset.

- Verify all data directories exists on startup. (to cover case user did
"rm -rf /var/lib/piserver" but forget to recreate directories afterwards)

- Show warning if running Pixel in persistence mode.

Fixes #1