dhcpcd: disable link detection
Fixes #20

Disable dhcpcd link detection to prevent it removing IP from
interface when link gets temporarily disconnected, which it is not
unable to recover from, due to it wanting to run scripts that are
located on the nfs share it is unable to access.

Disable wifi by default.
People network booting generally do not need it, and dhcpcd does
need link detection to know when wifi association has completed.