Fix mouse cursor on startup
Because we run under sudo not all environment variables are passed
to us including one that gdk_notify_startup_complete() uses when
notifying the window manager that we finished starting up.

Because the window manager does not receive notification the mouse
cursor remains a hourglass wait cursor for a while, even though
we finished starting up, until it eventually times out.

Set StartupNotify=false in .desktop file as workaround.