Fix setting /etc/nslcd.conf when using external LDAP/AD server
When an external server is used the piserver.pem TLS
certificate is not being generated.
The post-installation script script only altered /etc/nslcd.conf
if that did exists, which is the wrong behavior.
(Escaped testing because if you re-install piserver multiple times
there may still exist a piserver.pem from a previous installation
that did not use external server)

Ref #72