Warn user if STP is enabled on switch
The Spanning-Tree-Protocol which is enabled by default on Cisco switches
conflicts with Pi 3 network booting, due to it taking 30+ seconds
before it is done testing for network loops, and moves to forwarding
state in which it lets normal network traffic through.
The Pi 3 network boot firmware does not try that long.

Try to detect if STP is enabled by listening for BDPU packets
and warn user about it if it is, so they do not have to wonder
why adding Pi to Piserver does not work.

(Limitation:

STP detection only works if piserver computer is connected directly
through the problem switch. And not when running Piserver under
Vmware/Virtualbox as BDPU packets can only be seen on the direct link,
and are not forwarded over other (real or virtual) switches).

Ref: https://www.raspberrypi.org/forums/viewtopic.php?f=36&t=246122
("Piserver in a school not finding 25* RP3B+")